Check node num length before making a new node info or user core data object
