+++ Meshtastic/Extensions/CoreData/UserEntityExtension.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import CoreData
 
 extension UserEntity {
 	
@@ -26,4 +27,17 @@ extension UserEntity {
 		let unreadMessages = messageList.filter{ ($0 as AnyObject).read == false } 
 		return unreadMessages.count
 	}
+}
+
+
+public func createUser(num: Int64, context: NSManagedObjectContext) -> UserEntity {
+	let newUser = UserEntity(context: context)
+	newUser.num = Int64(num)
+	let userId = String(format:"%2X", num)
+	newUser.userId = "!\(userId)"
+	let last4 = String(userId.suffix(4))
+	newUser.longName = "Meshtastic \(last4)"
+	newUser.shortName = last4
+	newUser.hwModel = "UNSET"
+	return newUser
 }
